separate indicators for track vs collections
