+++ packages/mobile/src/components/offline-downloads/DownloadStatusIndicator.tsx
@@ -1,74 +1,91 @@
-import { useMemo } from 'react'
+import type { Nullable } from '@audius/common'
+import type { StyleProp, ViewStyle } from 'react-native'
+import { View } from 'react-native'
+import Rive from 'rive-react-native'
 
-import { cacheCollectionsSelectors } from '@audius/common'
-import { useSelector } from 'react-redux'
-
-import { useIsOfflineModeEnabled } from 'app/hooks/useIsOfflineModeEnabled'
-import {
-  getIsCollectionMarkedForDownload,
-  getIsAnyDownloadInProgress,
-  getTrackOfflineDownloadStatus,
-  getIsAllDownloadsErrored
-} from 'app/store/offline-downloads/selectors'
+import IconDownloadFailed from 'app/assets/images/iconDownloadFailed.svg'
+import IconDownloadInactive from 'app/assets/images/iconDownloadInactive.svg'
+import IconDownloadQueued from 'app/assets/images/iconDownloadQueued.svg'
+import IconDownloaded from 'app/assets/images/iconDownloaded.svg'
 import { OfflineDownloadStatus } from 'app/store/offline-downloads/slice'
+import { makeStyles } from 'app/styles'
+import { useThemeVariant } from 'app/utils/theme'
 
-import { DownloadStatusIndicator as DownloadStatusIndicatorBase } from './DownloadStatusIndicatorBase'
-
-const { getCollection } = cacheCollectionsSelectors
-
-type TrackDownloadIndicatorProps = {
-  trackId?: number
-  collectionId?: string
+type DownloadStatusIndicatorProps = {
+  status: Nullable<OfflineDownloadStatus>
   size?: number
+  style?: StyleProp<ViewStyle>
 }
 
-export const DownloadStatusIndicator = ({
-  collectionId,
-  trackId,
-  size = 24
-}: TrackDownloadIndicatorProps) => {
-  const isOfflineModeEnabled = useIsOfflineModeEnabled()
-  const isMarkedForDownload = useSelector(
-    getIsCollectionMarkedForDownload(collectionId)
-  )
+const useStyles = makeStyles(({ palette }) => ({
+  iconDownloadQueued: {
+    fill: palette.neutralLight4
+  },
+  iconDownloaded: {
+    fill: palette.secondary
+  },
+  iconDownloadFailed: {
+    fill: palette.secondary
+  },
+  iconDownloadInactive: {
+    fill: palette.neutralLight4
+  }
+}))
 
-  const trackDownloadStatus = useSelector(
-    getTrackOfflineDownloadStatus(trackId)
-  )
+export const DownloadStatusIndicator = (
+  props: DownloadStatusIndicatorProps
+) => {
+  const { status, size = 24, style } = props
+  const styles = useStyles()
+  const themeVariant = useThemeVariant()
 
-  const collection = useSelector((state) =>
-    getCollection(state, {
-      id: isMarkedForDownload && collectionId ? parseInt(collectionId) : null
-    })
-  )
-
-  const trackIds = useMemo(() => {
-    return (
-      collection?.playlist_contents?.track_ids?.map(
-        (trackData) => trackData.track
-      ) ?? []
-    )
-  }, [collection])
-
-  const isAnyDownloadInProgress = useSelector((state) =>
-    getIsAnyDownloadInProgress(state, trackIds)
-  )
-
-  const isAllDownloadsErrored = useSelector((state) =>
-    getIsAllDownloadsErrored(state, trackIds)
-  )
-
-  const downloadStatus =
-    trackDownloadStatus ??
-    (isMarkedForDownload
-      ? isAnyDownloadInProgress
-        ? OfflineDownloadStatus.LOADING
-        : isAllDownloadsErrored
-        ? OfflineDownloadStatus.ERROR
-        : OfflineDownloadStatus.SUCCESS
-      : null)
-
-  if (!isOfflineModeEnabled) return null
-
-  return <DownloadStatusIndicatorBase status={downloadStatus} size={size} />
+  const renderIndicator = () => {
+    switch (status) {
+      case OfflineDownloadStatus.INIT:
+        return (
+          <IconDownloadQueued
+            fill={styles.iconDownloadQueued.fill}
+            height={size}
+            width={size}
+          />
+        )
+      case OfflineDownloadStatus.LOADING:
+        return (
+          <View>
+            <Rive
+              style={{ height: size, width: size }}
+              resourceName={`downloading_${themeVariant}`}
+              autoplay
+            />
+          </View>
+        )
+      case OfflineDownloadStatus.SUCCESS:
+        return (
+          <IconDownloaded
+            fill={styles.iconDownloaded.fill}
+            height={size}
+            width={size}
+          />
+        )
+      case OfflineDownloadStatus.ERROR:
+        return (
+          <IconDownloadFailed
+            fill={styles.iconDownloadFailed.fill}
+            height={size}
+            width={size}
+          />
+        )
+      case OfflineDownloadStatus.INACTIVE:
+        return (
+          <IconDownloadInactive
+            fill={styles.iconDownloadInactive.fill}
+            height={size}
+            width={size}
+          />
+        )
+      default:
+        return null
+    }
+  }
+  return <View style={style}>{renderIndicator()}</View>
 }
